Make RustStream interop type available in all configurations
