+++ skia-bindings/src/bindings.cpp
@@ -2753,3 +2753,117 @@ extern "C" SkOrderedFontMgr* C_SkOrderedFontMgr_new() {
 extern "C" void C_SkOrderedFontMgr_append(SkOrderedFontMgr* self, SkFontMgr* fontMgr) {
     self->append(sp(fontMgr));
 }
+
+//
+// SkStream <-> RustStream interop
+//
+
+class RustStream : public SkStream {
+    void *m_data;
+    size_t m_length;
+    bool m_isEof;
+
+    size_t (*m_read)(void *, void *, size_t);
+
+    bool (*m_seekAbsolute)(void *, size_t);
+
+    bool (*m_seekRelative)(void *, long);
+
+public:
+    RustStream(
+            void *data,
+            size_t length,
+            size_t (*read)(void *, void *, size_t),
+            bool (*seekAbsolute)(void *, size_t),
+            bool (*seekRelative)(void *, long)
+    );
+
+    size_t read(void *buffer, size_t count);
+
+    bool rewind();
+
+    bool seek(size_t pos);
+
+    bool move(long offset);
+
+    bool isAtEnd() const;
+
+    bool hasLength() const;
+
+    size_t getLength() const;
+};
+
+RustStream::RustStream(
+        void *data,
+        size_t length,
+        size_t (*read)(void *, void *, size_t),
+        bool (*seekAbsolute)(void *, size_t),
+        bool (*seekRelative)(void *, long)
+) :
+        m_data(data),
+        m_length(length),
+        m_isEof(false),
+        m_read(read),
+        m_seekAbsolute(seekAbsolute),
+        m_seekRelative(seekRelative) {}
+
+size_t RustStream::read(void *buffer, size_t count) {
+    if (this->m_isEof) return 0;
+
+    size_t out = (this->m_read)(this->m_data, buffer, count);
+
+    if (!out) {
+        this->m_isEof = true;
+    }
+
+    return out;
+}
+
+bool RustStream::rewind() {
+    return this->seek(0);
+}
+
+bool RustStream::seek(size_t pos) {
+    if (this->m_seekAbsolute) {
+        return (this->m_seekAbsolute)(this->m_data, pos);
+    } else {
+        return false;
+    }
+}
+
+bool RustStream::move(long offset) {
+    if (this->m_seekRelative) {
+        return (this->m_seekRelative)(this->m_data, offset);
+    } else {
+        return false;
+    }
+}
+
+bool RustStream::hasLength() const {
+    return this->m_length != (size_t) - 1;
+}
+
+size_t RustStream::getLength() const {
+    return this->m_length;
+}
+
+bool RustStream::isAtEnd() const {
+    return this->m_isEof;
+}
+
+extern "C" void C_RustStream_construct(
+        RustStream *out,
+        void *data,
+        size_t length,
+        size_t (*read)(void *, void *, size_t),
+        bool (*seekAbsolute)(void *, size_t),
+        bool (*seekRelative)(void *, long)
+) {
+    new(out) RustStream(data, length, read, seekAbsolute, seekRelative);
+}
+
+extern "C" void C_RustStream_destruct(
+    RustStream *stream
+) {
+    stream->~RustStream();
+}
