+++ skia-safe/src/interop/stream.rs
@@ -3,13 +3,10 @@
 //!
 //! Bindings that wrap functions that use Skia stream types, _must_ use Rust streams instead.
 
-use crate::prelude::*;
-use crate::Data;
+use crate::{prelude::*, Data};
 use skia_bindings as sb;
 use skia_bindings::{SkDynamicMemoryWStream, SkMemoryStream, SkStream, SkStreamAsset, SkWStream};
-#[cfg(feature = "svg")]
-use std::{ffi, io};
-use std::{fmt, marker::PhantomData, ptr};
+use std::{ffi, fmt, io, marker::PhantomData, ptr};
 
 /// Trait representing an Skia allocated Stream type with a base class of SkStream.
 #[repr(transparent)]
@@ -158,30 +155,25 @@ impl DynamicMemoryWStream {
     }
 }
 
-#[cfg(feature = "svg")]
 pub struct RustStream<'a> {
     inner: Handle<sb::RustStream>,
     _phantom: PhantomData<&'a mut ()>,
 }
 
-#[cfg(feature = "svg")]
 impl RustStream<'_> {
     pub fn stream_mut(&mut self) -> &mut SkStream {
         self.inner.native_mut().base_mut()
     }
 }
 
-#[cfg(feature = "svg")]
 impl NativeBase<SkStream> for sb::RustStream {}
 
-#[cfg(feature = "svg")]
 impl NativeDrop for sb::RustStream {
     fn drop(&mut self) {
         unsafe { sb::C_RustStream_destruct(self) }
     }
 }
 
-#[cfg(feature = "svg")]
 impl<'a> RustStream<'a> {
     pub fn new<T: io::Read>(val: &'a mut T) -> Self {
         unsafe extern "C" fn read_trampoline<T>(
